Move GetZoneKey function to component-helpers
