+++ pkg/util/node/node.go
@@ -174,42 +174,6 @@ func GetNodeIP(client clientset.Interface, name string) net.IP {
 	return nodeIP
 }
 
-// GetZoneKey is a helper function that builds a string identifier that is unique per failure-zone;
-// it returns empty-string for no zone.
-// Since there are currently two separate zone keys:
-//   * "failure-domain.beta.kubernetes.io/zone"
-//   * "topology.kubernetes.io/zone"
-// GetZoneKey will first check failure-domain.beta.kubernetes.io/zone and if not exists, will then check
-// topology.kubernetes.io/zone
-func GetZoneKey(node *v1.Node) string {
-	labels := node.Labels
-	if labels == nil {
-		return ""
-	}
-
-	// TODO: "failure-domain.beta..." names are deprecated, but will
-	// stick around a long time due to existing on old extant objects like PVs.
-	// Maybe one day we can stop considering them (see #88493).
-	zone, ok := labels[v1.LabelFailureDomainBetaZone]
-	if !ok {
-		zone, _ = labels[v1.LabelTopologyZone]
-	}
-
-	region, ok := labels[v1.LabelFailureDomainBetaRegion]
-	if !ok {
-		region, _ = labels[v1.LabelTopologyRegion]
-	}
-
-	if region == "" && zone == "" {
-		return ""
-	}
-
-	// We include the null character just in case region or failureDomain has a colon
-	// (We do assume there's no null characters in a region or failureDomain)
-	// As a nice side-benefit, the null character is not printed by fmt.Print or glog
-	return region + ":\x00:" + zone
-}
-
 type nodeForConditionPatch struct {
 	Status nodeStatusForPatch `json:"status"`
 }
