+++ pkg/util/node/node_test.go
@@ -266,84 +266,3 @@ func TestGetHostname(t *testing.T) {
 
 	}
 }
-
-func Test_GetZoneKey(t *testing.T) {
-	tests := []struct {
-		name string
-		node *v1.Node
-		zone string
-	}{
-		{
-			name: "has no zone or region keys",
-			node: &v1.Node{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{},
-				},
-			},
-			zone: "",
-		},
-		{
-			name: "has beta zone and region keys",
-			node: &v1.Node{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						v1.LabelFailureDomainBetaZone:   "zone1",
-						v1.LabelFailureDomainBetaRegion: "region1",
-					},
-				},
-			},
-			zone: "region1:\x00:zone1",
-		},
-		{
-			name: "has GA zone and region keys",
-			node: &v1.Node{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						v1.LabelTopologyZone:   "zone1",
-						v1.LabelTopologyRegion: "region1",
-					},
-				},
-			},
-			zone: "region1:\x00:zone1",
-		},
-		{
-			name: "has both beta and GA zone and region keys",
-			node: &v1.Node{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						v1.LabelTopologyZone:            "zone1",
-						v1.LabelTopologyRegion:          "region1",
-						v1.LabelFailureDomainBetaZone:   "zone1",
-						v1.LabelFailureDomainBetaRegion: "region1",
-					},
-				},
-			},
-			zone: "region1:\x00:zone1",
-		},
-		{
-			name: "has both beta and GA zone and region keys, beta labels take precedent",
-			node: &v1.Node{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						v1.LabelTopologyZone:            "zone1",
-						v1.LabelTopologyRegion:          "region1",
-						v1.LabelFailureDomainBetaZone:   "zone2",
-						v1.LabelFailureDomainBetaRegion: "region2",
-					},
-				},
-			},
-			zone: "region2:\x00:zone2",
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			zone := GetZoneKey(test.node)
-			if zone != test.zone {
-				t.Logf("actual zone key: %q", zone)
-				t.Logf("expected zone key: %q", test.zone)
-				t.Errorf("unexpected zone key")
-			}
-		})
-	}
-}
