+++ staging/src/k8s.io/component-helpers/node/OWNERS
@@ -1,0 +1,8 @@
+# See the OWNERS docs at https://go.k8s.io/owners
+
+approvers:
+- sig-node-approvers
+reviewers:
+- sig-node-reviewers
+labels:
+- sig/node
